requirements.txt
torch
torchvision
gym
stable-baselines3[extra]
plotly
keras
pytest-shutil
openai-clip
transformers
accelerate
diffusers
datasets
opencv-python
scipy
kaggle
matplotlib
nltk
numpy
pandas
Pattern
sacrebleu
pillow
tqdm